Make prob025-polettix.pl of the Euler problem set work again
Basically, it was necessary to copy the printable() routine from
euler/prob063-polettix.pl and use it to return the appropriate value for the
most significant part of the number (the algorithm uses a work around which
was necessary before BigInts were implemented) such that the current length
could be calculated.  Long story short, it works again.