[wsg] add *some* content to test3.txt input file
The information about the Winter Scripting Games is only available on
web.archive.org (the original info at Microsoft is long gone) thus it's not
a major problem that the content of this input file differs from that which
can't be found anymore (unfortunately).  The example code is really what
matters here and the example's output is now testable after this change.