Remove order depencence on output of grph example
The problem text on the Rosalind Project web site specifies that the edges
can be returned in any order.  Hence it is important that the correct edges
are returned, not their ordering.  Thus one can compare a sorted lists of
edges to ensure that the correct elements are present and that the correct
answer has been generated.