[euler] enable loop param to be writable in loop
The code `for $file.IO.lines[^$lines] <-> $line` iterates over a lazy list
of lines without a container, thus it gives the error `Cannot assign to an
immutable value` (since `$line` is immutable in this case).  To enable
`$line` to be writable within the loop, one needs to replace `<-> $line`
with `-> $line is copy`.  Thanks to jnthn++ for the clear explanation.
Unfortunately, the script doesn't produce the correct answer, but that's
another issue.