[99probs] fix P34-rhebus post GLR
It turns out that `[+](True)` returns `True` (as spec-ed), however this is
exactly what `2 <<coprime<< list(1..^2)` generates, and since it's not
numeric, it doesn't have the correct return type.  The solution is to
explicitly cast the result of checking for a greatest common divisor for
being 1 to a numerical value and thereafter the `[+]` reduction operator
will do the right thing for all values passed in to the general case of the
totient function.