Simplify calculation of places to rotate
The previous calculation used a ternary (?? !!) expression to decide whether
or not to calculat the value of the `places` variable.  This had the "or"
part of the ternary conditional operator use a null sink and hence this line
could simply be replaced by a postfix `if`.

This removes the "Useless use of () in sink context" warning we were
getting.