Fix sporadic categories test failure
We were getting the following error in the categories tests:

```
    not ok 1 - Appended subcategories are a Categories object
    # Failed test 'Appended subcategories are a Categories object'
    # at t/003-categories.t line 60
No such method 'categories-list' for invocant of type 'Any'
  in block <unit> at t/003-categories.t line 63
```

which seemed to be happening because the code was calling a method on an
atrribute on the first element of an array.  An initial investigation
showed that the subcategories attribute was actually `Any` and not a
`Categories` (as had been tested in the previous line of code, buh?)
Nevertheless, after debug code was added to print the value of the first
element of the array, the test failures disappeared, hence it looked
like something was possibly not being set or created correctly.  By
setting the first element of the categories-list to a temporary
variable, we were able to avoid the subcategories attribute being set to
`Any` and hence the `categories-list` method could be found and the
tests pass again.