Install test-required dependencies
The `--/test` option to `zef` doesn't install the dependencies needed to
run all of the tests, hence we need to use `--test`.  However, in order
to avoid running the tests for all of the other dependencies, we first
install the base dependencies *without* tests and then install the
remainder *with* tests.  This way dependencies required for running the
tests (such as `Test::META`) are installed and the test suite passes.